feat: tweaks to how we configure logging (#257)
* feat: tweaks to how we configure logging

This changes the signatures that we use to configure logging
for the client.  Prior to this commit the constructors were
taking a vararg array of loggerFactories, which might lead
users to think they could pass more than one.

This commit adds alternate constructors that accept a single
logger factory instead.