feat: More tweaks to logging API
As I was working on example code for consuming the logging API
I found a few other pieces of it that weren't structured quite
right.

I now have a working example of using the logging API with logrus,
which I will file a separate PR for, but it relies on these
logging changes.