chore: simplify logic of batchutils functions
This commit gets rid of a goroutine in the batch utils functions
that is arguably unnecessary.

When we are populating the worker input channels, we were restricting
the size of the channel to be only as large as the max number of
workers. Then we needed a producer goroutine to feed the keys into
that channel as the workers removed them.

However, since the response channel is already sized to handle the
entire result set, and the items we put into the input channel are
already in memory, sizing the channel so that it is large enough
to hold pointers to all of them seems like a reasonable thing to
do. It shouldn't increase memory usage much, and it simplifies the
code considerably. We get rid of a goroutine and a context that we
would otherwise need to be careful about leaking.