fix: Allow subscription to complete when cancelled
If the context that was used to create a subscription is cancelled,
it becomes visible inside the subscription's `Item` function via
`cancelContext.Done()`. However, we were not checking for that
condition when the grpc client's `RecvMsg` function returned,
so after a cancel, we were reconnecting the stream. This made it
impossible to write a goroutine that called this Item function
and then cancel it successfully for a clean shutdown.

This commit adds in some extra logging and an additional case
statement to allow the subscription Item function to exit
when the cancelContext is done.