fix: don't log error message when subscription context cancelled (#416)
Prior to this commit, when a topic subscription was cancelled,
we would log an error message stating that we were going to attempt
to reconnect even when we aren't actually going to.

This commit moves the error log statement so that it only happens
if the contexts are not Done. This prevents users from seeing
errant error log messages when they are performing a clean shutdown.