chore: add example for ShardedTopicClient (#417)
* chore: add example for ShardedTopicClient

This commit adds a new example in the examples dir; it contains
a drop-in-replacement for the TopicClient that will allow you
to specify a number of shards to split your topic into. Behind
the scenes it will create multiple topics and split up your
messages between them, which can help distribute load more
evenly and improve performance for extremely high-volume topics.
It conforms to the existing TopicClient interface, so it can
be consumed without any code changes beyond the constructor.