fix: use sync.Map to track first-time agent headers
Prior to this commit we were using a regular `map` object for
keeping track of whether the first-time headers have been
sent for a given client. These `map` objects are not safe for
concurrent access, and will cause the program to crash if there
is a concurrent read while a write is occurring.

This commit switches to a `sync.Map` which is safe for concurrent
access.