test: identify services under test w/labels and setup conditionally
Modifies the tests to use labels to identify each service under
test. We then update the makefile to call ginkgo with a label filter,
and can then programmatically detect if (for example) the storage
service is under test. When the storage service is under test, we set
up the test resources. When the storage service is not under test, we
do not do that.

Previously we always stood up storage service test resources, which
caused errors in places where the storage service was not deployed
even if we weren't specifically testing the storage service.