fix: interpret get rank response when sorted set found
Previously we did not interpret an element hit vs miss when the sorted
set was found on a sorted set get rank request. In the event a user
queried the rank of an element not present in an existing sorted set,
a rank 0 would be returned.

To correct this we inspect the `result` field, check hit vs miss, and
return a `SortedSetGetRankMiss` on a miss.