feat: document all rpc's and if they are idempotent or not (#558)
Based on the current cache service proto, we add in all RPCs. For
maintainability, we arrange them in the same order as in the proto
file. For the RPCs that are new relative to when this was last
updated, add a comment and mark whether they are retryable or not.