chore: remove grpcResponse from request structs
Rather than having the requestor store the grpc responses as state, we are now just passing the response in the data client returned from makeRequest into interpretGrpcResponse. In the vast majority of cases, this was the only reason that field existed. This also allows the response to be modified by the middleware layer, which was previously not possible. In the small number of cases where the grpc response was used outside of the data client, we retain the field in the request struct.