feat: add base and gzip compression middleware for scalar get and set requests (#628)
Work towards momentohq/dev-eco-issue-tracker#1219

Creates a CompressionMiddleware interface that can accept a CompressionStrategy for gzip or zstd.
The gzip version is folded into the config/middleware/impl folder since gzip is part of the go standard library.
The zstd version will be published as a different module named client-sdk-go-compression-zstd (mirroring the JS version) so that users can opt into the zstd dependency.